Combined Add Date button and form into one reusable component. #7
