+++ server/public/styles.css
@@ -404,6 +404,7 @@ body {
   background-color: rgb(0, 0, 0, 0.5);
   min-height: 100vh;
   top: 0;
+  left: 0;
   z-index: 1;
 }
